implement download and upload feature

+++ Phase10_Drive/src/main/java/ir/shelmos/drive/repositories/PhotosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ir.shelmos.drive.models.Photo;
 
-public interface PhotosRepository extends CrudRepository<Photo, Long> {
+public interface PhotosRepository extends CrudRepository<Photo, Integer> {
 
 }
